Code review comments; one cases.js file only; Tree click opens a new case in the same tab.
