--- katana/config.json
@@ -1,11 +1,0 @@
-{
-    "engineer": "",
-    "idfdir": "",
-    "projdir": "/home/keen/Desktop/demo4/warriorframework/wftests/warrior_tests/projects",
-    "pythonpath": "",
-    "pythonsrcdir": "/home/keen/Desktop/demo4/warriorframework/warrior",
-    "testdata": "",
-    "testsuitedir": "/home/keen/Desktop/demo4/warriorframework/wftests/warrior_tests/suites",
-    "warhorn_config": "",
-    "xmldir": "/home/keen/Desktop/demo4/warriorframework/wftests/warrior_tests/testcases"
-}
