+++ katana/native/cases/templates/listAllCases.html
@@ -3,33 +3,14 @@
 
 <body>
 
-<div class="tool-bar">
-<button id="addNewCase" class="btn btn-success" onclick="cases.startNewCase()">Add New Testcase</button>
-</div>
-<!--
-<div class="tool-bar"><div katana-click="settings.closeSetting" class="button close">Close</div><div katana-click="settings.save" class="button save">Save</div></div>
-<div class="flex-rows to-save" auto-init="settings.emailSettings.init" post-url="settings/jira_setting_handler">
--->
+<div id="OverwriteCaseHere">
+
 <div class="csrf-container">{% csrf_token %}</div>
 
-
 <h2>{{title}}</h2>
-<div id="myCaseTree"></div>
+<div id="myCaseTree" auto-init="cases.init"></div>
 <div id="mmm" class="hidden">{{treejs}}</div>
-
-<script> 
-
-$(".linkToCase").click(function() { 
-	var $elem = $(this);
-	var thePage = $elem.attr('href');
-	var theID   = $elem.attr('fname');
-    katana.$view.one('tabAdded', function(){
-   		caseApp.mapFullCaseJson(theID);
-    });
-	//
-	katana.templateAPI.load( thePage, null, null, 'case') ;
-});
-cases.displayTreeOfCases();
-</script>
+<button id="addNewCase" class="btn btn-success" onclick="cases.startNewCase()">Add New Testcase</button>
+</div>
 </body>
