+++ warrior/Tools/w_settings.xml
@@ -1,19 +1,14 @@
 <Default>
     <Setting name="def_dir">
-        <Logsdir></Logsdir> <!--Optional, uses default, support absolute path, relative path and environment variable(${ENV.DEF_LOGSDIR})-->
-        <Resultsdir></Resultsdir> <!--Optional, uses default, support absolute path, relative path and environment variable(${ENV.DEF_RESULTSDIR})-->
+        <Logsdir />
+        <Resultsdir />
     </Setting>
-    <Setting name="mail_to" mail_on="">
-    <!-- 'mail_on' attribute is to specify 'when to send an email'. Supported values are 'per_execution, first_failure & every_failure'.
-            1. per_execution(default) - to send an email after the end of case/suite/project execution
-            2. first_failure - to send an email as soon as the first failure occurs during suite/project execution
-            3. every_failure - to send an email whenever a case fails during suite/project execution
-         When 'mail_on attribute' is not given or left empty, default value will be used for sending email.
-         It also accepts comma separated values, example mail_on="per_execution, first_failure" -->
-        <smtp_host></smtp_host> <!-- Optional, use FNC smtp host smtp.fnc.fujitsu.com to notify test case execution results -->
-        <sender></sender> <!-- Optional, use sender's email such as Warrior_PBO@fnc.fujitsu.com -->
-        <receiver></receiver> <!-- Optional, use receiver's email such as FNC.User@fnc.fujitsu.com -->
-        <subject></subject> <!-- Optional, use this as subject line such as WARRIOR , additional test case execution results Pass/Fail and Test Execution File shall be appended to this -->
+    <Setting mail_on="" name="mail_to">
+
+        <smtp_host>smtp.fnc.fujitsu.com</smtp_host>
+        <sender>noreply@us.fujitsu.com</sender>
+        <receiver>sathyamoorthy.radhakrishnan@us.fujitsu.com</receiver>
+        <subject />
     </Setting>
