--- warrior/Warriorspace/Config_files/Config_File_Readme.txt
@@ -1,5 +1,0 @@
-This is the directory to hold any configuration files related to your test.
-A sample warrior recomended testdata file to send commands to a cli session is available in this directory.
-For specific needs users can create their own format of configuration/testdata files and 
-parse them by creating custom libraries in Warrior Framework.
-
