--- warrior/Warriorspace/Suites/demo_suite.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" ?>
-<TestSuite>
-        <Details>
-                <Name>Demo_Suite</Name>
-                <Title>Demo_Suite</Title>
-                <Engineer>Warrior_user</Engineer>
-                <Date>03/17/2017</Date>
-                <Time>21:37:23</Time>
-                <type Max_Attempts="" Number_Attempts="" exectype="sequential_testcases"/>
-                <State>Released</State>
-                <default_onError action="next" />
-                <Resultsdir/>
-        </Details>
-        <Requirements>
-                <Requirement>Requirement-demo-001</Requirement>
-                <Requirement>Requirement-demo-002</Requirement>
-        </Requirements>
-        <Testcases>
-                <Testcase>
-                        <path>../Testcases/Demo_Cases/Demo_Test_Of_Inventory_Management_System.xml</path>
-                        <context>positive</context>
-                        <runtype>sequential_keywords</runtype>
-                        <onError action="next" value=""/>
-                        <impact>impact</impact>
-                        <InputDataFile/>
-                        <Execute ExecType="Yes">
-                                <Rule Condition="" Condvalue="" Else="next" Elsevalue=""/>
-                        </Execute>
-                        <runmode type="Standard" value=""/>
-                </Testcase>
-                <Testcase>
-                        <path>../Testcases/Demo_Cases/Demo_Warrior_Rest_Automation_Capabilities.xml</path>
-                        <context>positive</context>
-                        <runtype>sequential_keywords</runtype>
-                        <onError action="next" value=""/>
-                        <impact>impact</impact>
-                        <InputDataFile/>
-                        <Execute ExecType="Yes">
-                                <Rule Condition="" Condvalue="" Else="next" Elsevalue=""/>
-                        </Execute>
-                        <runmode type="Standard" value=""/>
-                </Testcase>
-                <Testcase>
-                        <path>../Testcases/Demo_Cases/Demo_Warrior_Web_Automation_With_Selenium.xml</path>
-                        <context>positive</context>
-                        <runtype>sequential_keywords</runtype>
-                        <impact>impact</impact>
-                        <InputDataFile/>
-                        <Execute ExecType="Yes">
-                                <Rule Condition="" Condvalue="" Else="next" Elsevalue=""/>
-                        </Execute>
-                        <runmode type="Standard" value=""/>
-                        <onError action="next" value=""/>
-                </Testcase>
-        </Testcases>
-</TestSuite>
