+++ warrior/plugins/plugin_robot_wrapper/bin/__init__.py
